Heart disease prediction
Welcome to the Heart Disease Prediction repository powered by Logistic Regression! This project is dedicated to building an efficient predictive model for assessing the likelihood of heart diseases. Dive into the world of machine learning and explore the nuances of Logistic Regression for accurate risk assessment. Discover how to preprocess, train, and fine-tune the model to deliver precise predictions and insights. Join us on this journey to leverage the power of Logistic Regression for proactive heart health management and early risk detection.